Add Lent logical working under process

+++ WinFormsApp/PersonalExpenseCreditTracker/BLLayer/Lent/LentBLL.cs
@@ -126,7 +126,7 @@ namespace BLLayer.Lent
             return dataTable;
         }
 
-        public static DataTable retriveDataByUserIdAndFilterIdAtBll(string spName, int userId, string paramName1, int paramId1, string paramName2, int paramId2)
+        public static DataTable retriveDataByUserIdAndFilterIdAtBll(string spName, int userId, string paramName1, Decimal paramId1, string paramName2, Decimal paramId2)
         {
             DataTable dataTable = new DataTable();
             dataTable = SqlHelper.retriveDataByUserIdAndFilterIdAtDal(spName,userId,paramName1,paramId1,paramName2,paramId2);
